+++ lib/widgets/home_widgets/DashboardsWidgets.dart
@@ -59,39 +59,6 @@ class _DiaryIndicatorsState extends State<DiaryIndicators> {
             )
           ],
         );
-
-        /*return StreamBuilder<DayModel?>(
-          stream: dayProvider.getSpecificDayStream(actualDay.id),
-          builder: (context, snapshot) {
-            if (snapshot.connectionState == ConnectionState.waiting)
-              return CircularProgressIndicator();
-            else if (snapshot.hasError)
-              return Container();
-            else {
-              final DayModel? day = snapshot.data;
-
-              if (day == null)
-                return Text('Día no encontrado');
-              // Actualiza tu interfaz de usuario con los datos del día específico
-              return Row(
-                children: [
-                  CaloriesIndicator(
-                    tooltipBehavior: _proteinsTooltipBehavior,
-                    actualValue: day.caloriesConsumed.toInt(),
-                    maximumValue: user.targetCalories,
-                    size: 200,
-                  ),
-                  MacronutrientsIndicator(
-                    tooltipBehavior: _macrosTooltipBehavior,
-                    totalProteins: day.proteinsConsumed,
-                    totalCarbs: day.carbsConsumed,
-                    totalFats: day.fatsConsumed,
-                  )
-                ],
-              ); 
-            }
-          },
-        );*/
       },
     );
   }
